did some small changes, started implementing python codegen which will be required for further building

+++ src/include/BitStream.h
@@ -333,7 +333,7 @@ protected:
   }
 
   void writeString(const char *str, size_t str_len) {
-    WriteCompressed((uint16_t) str_len);
+    WriteCompressed((uint32_t) str_len);
     if (str_len)
       Write(str, (uint32_t) str_len);
   }
